Move external secret store protos into their own module
